backend/backend-requirements.txt
Django==5.1.6
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyspark==3.5.1
boto3>=1.26.0
botocore>=1.29.0
openpyxl>=3.1.2
requests>=2.31.0
minio>=7.2.15
